requirements.txt
requests
pywhatkit
pyautogui
python-dotenv
unidecode
